requirements/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Uploads==0.2.1
Werkzeug==2.3.7

# AI/ML Libraries (Updated versions for compatibility)
tensorflow>=2.15.0
torch>=2.0.1
torchvision>=0.15.2
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Image Processing
albumentations>=1.3.0
imgaug>=0.4.0

# YOLO and Object Detection
ultralytics>=8.0.0
roboflow>=1.2.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-flask>=1.2.0
